requirements.txt
pytest==8.3.2
playwright~=1.46.0
pytest-playwright<=0.4
requests==2.32.3
Faker==19.13.0
python-dotenv==0.21.1
allure-pytest~=2.13.0
pkce~=1.0.3